Moved encryption_interceptor to interceptors package
